logging/nxscope: fix various warnings reported by CodeChekcer
fix various warnings reported by CodeChekcer for nxscope

+++ logging/nxscope/nxscope.c
@@ -1120,7 +1120,7 @@ int nxscope_recv(FAR struct nxscope_s *s)
 
   /* Handle frame */
 
-  ret = nxscope_recv_handle(s, frame.id, frame.dlen, frame.data);
+  ret = nxscope_recv_handle(s, frame.id, (uint16_t)frame.dlen, frame.data);
   if (ret < 0)
     {
       _err("ERROR: nxscope_recv_handle failed %d\n", ret);
